+++ src/traits.rs
@@ -3,6 +3,21 @@ use rand_core::CryptoRngCore;
 
 use crate::{generate_prime_with_rng, generate_safe_prime_with_rng, is_prime_with_rng, is_safe_prime_with_rng};
 
+/// A type producing sieves for random prime generation.
+pub trait SieveFactory {
+    /// The type of items returning by the sieves.
+    type Item;
+
+    /// The resulting sieve.
+    type Sieve: Iterator<Item = Self::Item>;
+
+    /// Makes a sieve given an RNG and the previous exhausted sieve (if any).
+    ///
+    /// Returning `None` signals that the prime generation should stop.
+    fn make_sieve(&mut self, rng: &mut impl CryptoRngCore, previous_sieve: Option<&Self::Sieve>)
+        -> Option<Self::Sieve>;
+}
+
 /// Provides a generic way to access methods for random prime number generation
 /// and primality checking, wrapping the standalone functions ([`is_prime_with_rng`] etc).
 pub trait RandomPrimeWithRng {